[#5] feat : BottomNav OpenStatus 처리 로직 추가
- 하단바를 띄워줄 path는 BOTTOM_NAV_PATH_INFO 에 추가해줍니다.
