+++ src/components/BottomNav/BottomNav.tsx
@@ -1,22 +1,19 @@
-import React from 'react'
+import React, { useMemo } from 'react'
+import { useLocation } from 'react-router-dom';
+import { BottomNavItemData } from './BottomNav.type';
+import { BOTTOM_NAV_ITEM_DATA, BOTTOM_NAV_PATH_INFO } from './BottomNav.util';
 import { BottomNavItemLink, BottomNavLabel, Root } from './styles'
 
-export type BottomNavItemData = {
-  path: string;
-  label: string;
-};
-
-export const BOTTOM_NAV_ITEM_DATA: Array<BottomNavItemData> = [
-  { path: '/', label: '홈' },
-  { path: '/community', label: '커뮤니티' },
-  { path: '/temp1', label: '임시1' },
-  { path: '/closet', label: '옷장' },
-  { path: '/temp2', label: '임시2' },
-]
-
 const BottomNav = () => {
+
+  const {pathname} = useLocation();
+
+  const isOpen = useMemo(() => {
+    return BOTTOM_NAV_PATH_INFO[pathname];
+  },[pathname]);
+
   return (
-    <Root>
+    <Root isOpen={isOpen}>
       {BOTTOM_NAV_ITEM_DATA.map((item: BottomNavItemData) => (
         <BottomNavItemLink key={item.path} to={item.path}>
           <BottomNavLabel>{item.label}</BottomNavLabel>
@@ -26,4 +23,4 @@ const BottomNav = () => {
   );
 }
 
-export default BottomNav
+export default BottomNav;
