+++ src/components/BottomNav/index.tsx
@@ -1,1 +1,3 @@
 export * from './BottomNav';
+export * from './BottomNav.type';
+export * from './BottomNav.util';
