+++ src/components/BottomNav/styles.tsx
@@ -1,15 +1,20 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-export const Root = styled.div`
+export const Root = styled.div<{ isOpen: boolean }>`
   z-index: 50;
   position: fixed;
-  bottom: 0;
+  bottom: ${(props) => (props.isOpen ? '0' : '-50px')};
   left: 0;
   right: 0;
+  height: 50px;
 
   display: flex;
   justify-content: space-between;
+
+  visibility: ${(props) => (props.isOpen ? 'visible' : 'hidden')};
+  opacity: ${(props) => (props.isOpen ? '1' : '0')};
+  transition: bottom 300ms ease-in-out;
 
   border: 1px red solid;
 `
